+++ src/player/ai/nn/Brain.h
@@ -7,12 +7,12 @@ public:
 	Brain(int numberOfHiddenLayers, int numberOfInput, int numberOfOutput, int wSize);
 	explicit Brain(std::vector<std::string>& lines);
 	~Brain();
-	double* decide(double data[]);
+	float* decide(float data[]);
 	short getOutputSize() const { return output->getNumberOfValues(); }
 
 private:
-	Eigen::MatrixXd multiply(Layer* current, Layer* prevLayer);
-	void setValues(Layer* layer, Eigen::MatrixXd& mult) const;
+	Eigen::MatrixXf multiply(Layer* current, Layer* prevLayer);
+	void setValues(Layer* layer, Eigen::MatrixXf& mult) const;
 	Layer* input;
 	Layer* output;
 	std::vector<Layer*> hidden;
