+++ src/player/ai/nn/Layer.cpp
@@ -3,28 +3,28 @@
 #include <iostream>
 
 Layer::Layer(short numberOfNodes, short prevSize): numberOfNodes(numberOfNodes), prevSize(prevSize) {
-	values = new double[numberOfNodes];
+	values = new float[numberOfNodes];
 	std::fill_n(values, numberOfNodes, 0);
 
-	bias = new double[numberOfNodes];
+	bias = new float[numberOfNodes];
 	for (int i = 0; i < numberOfNodes; i++) {
 		bias[i] = (static_cast<float>(rand()) / RAND_MAX - 0.5) * 2;
 	}
-	w = new double[numberOfNodes * prevSize];
+	w = new float[numberOfNodes * prevSize];
 	for (int i = 0; i < numberOfNodes * prevSize; i++) {
 		w[i] = (static_cast<float>(rand()) / RAND_MAX - 0.5) * 2;
 	}
 
 }
 
-Layer::Layer(std::vector<double>& w, std::vector<double>& bias) {
-	values = new double[bias.size()];
+Layer::Layer(std::vector<float>& w, std::vector<float>& bias) {
+	values = new float[bias.size()];
 	std::fill_n(values, numberOfNodes, 0);
 
-	this->bias = new double[bias.size()];
+	this->bias = new float[bias.size()];
 	std::copy(bias.begin(), bias.end(), this->bias);
 
-	this->w = new double[w.size()];
+	this->w = new float[w.size()];
 	std::copy(w.begin(), w.end(), this->w);
 	numberOfNodes = bias.size();
 	prevSize = w.size() / bias.size();
@@ -36,7 +36,7 @@ Layer::~Layer() {
 	delete []w;
 }
 
-void Layer::setValues(double data[]) {
+void Layer::setValues(float data[]) {
 	//TODO musi sie zgadzac rozmiar czy to jest sprawdzane?
 	std::copy(data, data + numberOfNodes, values);
 }
