+++ src/player/ai/nn/Layer.h
@@ -5,24 +5,24 @@ class Layer {
 
 public:
 	Layer(short numberOfNodes, short prevSize);
-	Layer(std::vector<double>& w, std::vector<double>& bias);
+	Layer(std::vector<float>& w, std::vector<float>& bias);
 	~Layer();
-	void setValues(double data[]);
-	double* getValues() const { return values; }
+	void setValues(float data[]);
+	float* getValues() const { return values; }
 	short getNumberOfValues() const { return numberOfNodes; }
 	short getPrevSize() const { return prevSize; }
-	double* getW() const { return w; }
+	float* getW() const { return w; }
 	short getWSize() const { return numberOfNodes * prevSize; }
-	double getBias(int i) const { return bias[i]; }
-	void setValueAt(int i, double newValue) { values[i] = newValue; }
+	float getBias(int i) const { return bias[i]; }
+	void setValueAt(int i, float newValue) { values[i] = newValue; }
 private:
 
 	short numberOfNodes;
 	short prevSize;
 
-	double* values;
-	double* bias;
+	float* values;
+	float* bias;
 
-	double* w;
+	float* w;
 
 };
